refactor(worker): rename lifecycle helper module
