+++ tracecat/agent/executor_worker.py
@@ -22,7 +22,7 @@ from tracecat.agent.runtime_services import (
 from tracecat.agent.worker import new_sandbox_runner
 from tracecat.dsl.client import get_temporal_client
 from tracecat.logger import logger
-from tracecat.temporal.shutdown import run_worker_entrypoint
+from tracecat.temporal.worker_lifecycle import run_worker_entrypoint
 
 if TYPE_CHECKING:
     from temporalio.client import Client
